+++ inst/testthat/helper_autotest.R
@@ -408,19 +408,38 @@ run_autotest = function(learner, N = 30L, exclude = NULL, predict_types = learne
 #'   run_paramtest(learner, fun, exclude)
 #'   expect_true(result, info = result$error)
 #' })
-run_paramtest = function(learner, fun, exclude = character()) {
-  par_learner = learner$param_set$ids()
-  par_package = formalArgs(fun)
+run_paramtest = function(learner, fun, exclude = character(), tag = NULL) {
+  par_learner = learner$param_set$ids(tags = tag)
+  if (checkmate::test_list(fun)) {
+    par_package = unlist(lapply(fun, formalArgs))
+  } else {
+    par_package = formalArgs(fun)
+  }
 
   missing = setdiff(par_package, par_learner)
   missing = setdiff(missing, c(exclude, "..."))
 
-  if (length(missing) == 0L)
+  extra = setdiff(par_learner, par_package)
+  extra = setdiff(extra, c(exclude, "..."))
+
+  if (length(c(missing, extra)) == 0L)
     return(TRUE)
 
-  error = sprintf("Missing parameters for learner '%s': %s",
-    learner$id, paste0(missing, collapse = ", "))
-  list(ok = FALSE, error = error, missing = missing)
+  merror = eerror = character(0)
+
+  if (length(missing) > 0) {
+    merror = sprintf("Missing parameters for learner '%s': %s",
+      learner$id, paste0(missing, collapse = ", "))
+  }
+
+  if (length(extra) > 0) {
+    eerror = sprintf("Extra parameters for learner '%s': %s",
+      learner$id, paste0(extra, collapse = ", "))
+  }
+
+  error = paste(merror, eerror, sep = "\n")
+
+  list(ok = FALSE, error = error, missing = missing, extra = extra)
 }
 
 # Helper function to convert a vector of probabilities to a matrix